Rename core.otel package to core.trace
The SpanListener mechanism is for all trace sources (OTLP, Zipkin,
Kafka), not OTLP-specific. The package name core.trace better
reflects this.
