--- oap-server/analyzer/gen-ai-analyzer/src/main/resources/META-INF/services/org.apache.skywalking.oap.server.core.otel.SpanListener
@@ -1,1 +1,0 @@
-org.apache.skywalking.oap.analyzer.genai.listener.GenAISpanListener
